server/requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
